Inject check job into CI workflow as ultimate flag
This patch adds a job that is able to accurately signal whether
all the expectations of the required jobs to succeed are met.

This job can then be used as a source of truth for judging whether
"CI passes" and can be used in the branch protection. It also plays
a role of a convenient "gate" — this is the only job that would have
to be listed in the branch protection as opposed to listing every
single job name generated by the test matrix (and they all have
different names — it's not possible to just select one `test` job
name).

Ref: https://github.com/re-actors/alls-green#why