Use the ruff formatter, instead of black
I have based this commit on:
- ruff-pre-commit README.md | Using Ruff with pre-commit
  https://github.com/astral-sh/ruff-pre-commit/blob/main/README.md
- The Ruff Formatter | Conflicting lint rules
  https://docs.astral.sh/ruff/formatter/#conflicting-lint-rules

Support for the ruff formatter was added to pytest-ruff by commits from
October 2023, released the same day as versions 0.2 and 0.2.1. Hence, it
makes sense to require pytest-ruff ≥ 0.2.1 now that we use the ruff
formatter.

Support for `quote-style = "preserve"` was added to ruff in the last
couple of weeks, therefore we require the latest version, ruff ≥ 0.1.8.
This option is equivalent to `skip-string-normalization` in black.